feat:[FE] 주문 관련 코드 리펙토링
